RF_viewTest
